// client/src/types/authTypes.ts
export enum AuthEvents {
  AUTH_INIT = "AUTH/AUTH_INIT",
  REGISTER_SUCCESS = "AUTH/REGISTER_SUCCESS",
  REGISTER_FAIL = "AUTH/REGISTER_FAIL",
  AUTH_SUCCESS = "AUTH/AUTH_SUCCESS",
  AUTH_DELETE = "AUTH/AUTH_DELETE",
  AUTH_FAIL = "AUTH/FAIL",
  LOGOUT = "AUTH/LOGOUT",
  PROFILE_UPDATE = "AUTH/PROFILE_UPDATE",
}